Better heap (force) flags clearing
The original method of a whitelist for the heap flags and simply zeroing the force flags is overly aggressive and also clears flags set by protectors (leading to detections) and flags set by the debuggee (leading to crashes due to missing e.g. HEAP_CREATE_ALIGN_16).
Switch to a blacklist-based approach of clearing only those (force) flags that are valid and indicate the presence of a debugger and/or NtGlobalFlags